server/requirements.txt
django==1.9
djangorestframework
markdown
django-filter  # Filtering support
drfdocs
